debug

+++ Application/Interfaces/Repositories/Base/IBaseRepository.cs
@@ -14,7 +14,7 @@ namespace Application.Interfaces.Repositories.Base
 
         Task<List<T>> GetListAsync();
 
-        Task<T> GetByIdAsync(int id);
+        Task<T> GetByIdAsync(long id);
 
         Task<long> InsertAsync(T entity);
 
@@ -36,7 +36,7 @@ namespace Application.Interfaces.Repositories.Base
 
         Task<List<T>> GetListAsync();
 
-        Task<T> GetByIdAsync(int id);
+        Task<T> GetByIdAsync(long id);
 
         Task<long> InsertAsync(T entity);
